+++ src/components/Home/List.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import DiaryItem from './DiaryItem';
+
+const diaries = [
+  { id: 1 },
+  { id: 2 },
+  { id: 3 },
+  { id: 4 },
+  { id: 5 },
+  { id: 6 },
+  { id: 7 },
+  { id: 8 },
+  { id: 9 },
+  { id: 10 },
+];
+
+const List = () => {
+  return (
+    <div>
+      {diaries.map((diary) => {
+        return <DiaryItem key={diary.id} />;
+      })}
+    </div>
+  );
+};
+
+export default List;
