+++ src/pages/Chat/Profile/Profile.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables/colors.scss';
-@import '../../styles/variables/fonts.scss';
+@import '../../../styles/variables/colors.scss';
+@import '../../../styles/variables/fonts.scss';
 
 .profileBefore {
   display: flex;
