--- src/pages/Detail/SelectTag.module.scss
@@ -1,2 +1,0 @@
-@import '../../styles/variables/colors.scss';
-@import '../../styles/variables/fonts.scss';
