+++ src/pages/Home/Home.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import useCalendar from '../../hooks/useCalendar';
 import styles from './Home.module.scss';
-import HomeCalendar from './HomeCalendar';
-import List from './List';
+import HomeCalendar from '../../components/Home/HomeCalendar';
+import List from '../../components/Home/List';
 import { ListIcon, DownChevron, Calendar32 } from '../../assets';
 import BottomNav from '../../components/BottomNav/BottomNav';
 import HomeHeader from '../../components/Headers/HomeHeader';
@@ -34,7 +34,7 @@ const Home = () => {
     <>
       <HomeHeader />
       <div className={styles.wholeWrapper}>
-      <HomeProfileHeader />
+        <HomeProfileHeader />
         <div className={styles.dateNav}>
           <div className={styles.currentDateBox}>
             <div className={styles.dateSelector} onClick={onSelectDate}>
@@ -71,7 +71,13 @@ const Home = () => {
           />
         )}
       </div>
-      {isSelectedDate ? <DateSelector clickOuter={setIsSelectedDate} isFullDate={false} isOpen={isSelectedDate} /> : null}
+      {isSelectedDate ? (
+        <DateSelector
+          clickOuter={setIsSelectedDate}
+          isFullDate={false}
+          isOpen={isSelectedDate}
+        />
+      ) : null}
       <BottomNav page={0} />
     </>
   );
