+++ src/pages/Tag/TagFilter/TagFilter.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables/fonts.scss';
-@import '../../styles/variables/colors.scss';
+@import '../../../styles/variables/fonts.scss';
+@import '../../../styles/variables/colors.scss';
 
 button {
   all: unset;
